GT_mmw/requirements.txt
absl-py==1.4.0
astor==0.8.1
cachetools==4.2.4
certifi==2021.5.30
cycler==0.11.0
dataclasses==0.8
gast==0.5.4
grpcio==1.48.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
nvidia-ml-py==11.525.150
nvitop==1.0.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
psutil==5.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
